requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
beautifulsoup4==4.12.3
multiprocess==0.70.16
pillow==10.3.0
python-slugify==8.0.4
requests==2.32.0